# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations API",
        "version": "0.1"
    },
    "paths": {
        "/api/organizations": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить список всех организаций",
                "operationId": "42efb13746ba71d73c330585f09ecef8",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить информацию об организации по ID",
                "operationId": "68e75650b204083886e125e02aaa2f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            }
        },
        "/api/organizations/building/{buildingId}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить список организаций в указанном здании",
                "operationId": "aae6fd413e5f76790a15e8b42e10919e",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity/{activityId}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить список организаций по виду деятельности",
                "operationId": "3445549c7daa790ee5f137b6309631d3",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/search/{name}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций по названию",
                "operationId": "1d54f372dba03097f781216b651ff59e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Название организации",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/radius": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций в заданном радиусе",
                "operationId": "d2b4e6612fb91764658089c856bd026b",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Широта точки центра",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Долгота точки центра",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска в километрах",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/searchActivityName/{activityName}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций по названию вида деятельности",
                "operationId": "05cbdb1e8a755de00e6bf681677af72b",
                "parameters": [
                    {
                        "name": "activityName",
                        "in": "path",
                        "description": "Название вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Организации",
            "description": "Организации"
        }
    ]
}